+++ examples/compute/instance/default_secgroup/base-firewall-datasource.tf
@@ -1,0 +1,4 @@
+data "vkcs_networking_secgroup" "default_secgroup" {
+  name = "default"
+  sdn  = "neutron"
+}