+++ CHANGELOG.md
@@ -8,6 +8,7 @@ description: |-
 # VKCS Provider's changelog
 
 #### v0.7.3 (unreleased)
+- Add security_group_ids argument to vkcs_compute_instance resource
 - Add ability to filter by extra_specs attribute in vkcs_compute_flavor data source.
 - Enhancement searching the closest appropriate flavor by min_ram argument in vkcs_compute_flavor data source.